Update and rename TextMenu.java to Menu.java
